requirements.txt
aiogram
feedparser
python-dotenv
pymysql
beautifulsoup4
playwright
pytz
tenacity
